+++ lib/src/results.dart
@@ -18,6 +18,8 @@
 import 'dart:async';
 import 'dart:ffi';
 
+import 'package:cancellation_token/cancellation_token.dart';
+
 import 'collections.dart';
 import 'native/realm_core.dart';
 import 'realm_class.dart';
@@ -162,6 +164,98 @@ extension RealmResultsOfObject<T extends RealmObjectBase> on RealmResults<T> {
   }
 }
 
+class _SubscribedRealmResult<T extends RealmObject> extends RealmResults<T> {
+  final String? subscriptionName;
+
+  _SubscribedRealmResult._(RealmResults<RealmObject> results, {this.subscriptionName})
+      : super._(
+          results.handle,
+          results.realm,
+          results.metadata,
+        );
+}
+
+extension RealmResultsOfRealmObject<T extends RealmObject> on RealmResults<T> {
+  /// Adds this [RealmResults] query to the set of active subscriptions.
+  /// The query will be joined via an OR statement with any existing queries for the same type.
+  ///
+  /// If a [name] is given this allows you to later refer to the subscription by name,
+  /// e.g. when calling [MutableSubscriptionSet.removeByName].
+  ///
+  /// If [update] is specified to `true`, then any existing query
+  /// with the same name will be replaced.
+  /// Otherwise a [RealmException] is thrown, in case of duplicates.
+  ///
+  /// [WaitForSyncMode] specifies how to wait or not wait for subscribed objects to be downloaded.
+  /// The default value is [WaitForSyncMode.firstTime].
+  ///
+  /// The [cancellationToken] is optional and can be used to cancel
+  /// the waiting for objects to be downloaded.
+  /// If the operation is cancelled, a [CancelledException] is thrown and the download
+  /// continues in the background.
+  /// In case of using [TimeoutCancellationToken] and the time limit is exceeded,
+  /// a [TimeoutException] is thrown and the download continues in the background.
+  ///
+  /// {@category Sync}
+  Future<RealmResults<T>> subscribe({
+    String? name,
+    WaitForSyncMode waitForSyncMode = WaitForSyncMode.firstTime,
+    CancellationToken? cancellationToken,
+    bool update = false,
+  }) async {
+    Subscription? existingSubscription = name == null ? realm.subscriptions.find(this) : realm.subscriptions.findByName(name);
+    late Subscription updatedSubscription;
+    realm.subscriptions.update((mutableSubscriptions) {
+      updatedSubscription = mutableSubscriptions.add(this, name: name, update: update);
+    });
+    bool shouldWait = waitForSyncMode == WaitForSyncMode.always ||
+        (waitForSyncMode == WaitForSyncMode.firstTime && subscriptionIsChanged(existingSubscription, updatedSubscription));
+
+    return await CancellableFuture.from<RealmResults<T>>(() async {
+      if (cancellationToken != null && cancellationToken.isCancelled) {
+        throw cancellationToken.exception!;
+      }
+      if (shouldWait) {
+        await realm.subscriptions.waitForSynchronization(cancellationToken);
+        await realm.syncSession.waitForDownload(cancellationToken);
+      }
+      return _SubscribedRealmResult._(this, subscriptionName: name);
+    }, cancellationToken);
+  }
+
+  /// Unsubscribe from this query result. It returns immediately
+  /// without waiting for synchronization.
+  ///
+  /// If the subscription is unnamed, the subscription matching
+  /// the query will be removed.
+  /// Return `false` if the [RealmResults] is not created by [subscribe].
+  ///
+  /// {@category Sync}
+  bool unsubscribe() {
+    bool unsubscribed = false;
+    if (realm.config is! FlexibleSyncConfiguration) {
+      throw RealmError('unsubscribe is only allowed on Realms opened with a FlexibleSyncConfiguration');
+    }
+    if (this is _SubscribedRealmResult<T>) {
+      final subscriptionName = (this as _SubscribedRealmResult<T>).subscriptionName;
+      realm.subscriptions.update((mutableSubscriptions) {
+        if (subscriptionName != null) {
+          unsubscribed = mutableSubscriptions.removeByName(subscriptionName);
+        } else {
+          unsubscribed = mutableSubscriptions.removeByQuery(this);
+        }
+      });
+    }
+    return unsubscribed;
+  }
+
+  bool subscriptionIsChanged(Subscription? existingSubscription, Subscription updatedSubscription) {
+    return existingSubscription == null ||
+        existingSubscription.objectClassName != updatedSubscription.objectClassName ||
+        existingSubscription.queryString != updatedSubscription.queryString;
+  }
+}
+
 /// @nodoc
 //RealmResults package internal members
 extension RealmResultsInternal on RealmResults {
@@ -180,12 +274,7 @@ extension RealmResultsInternal on RealmResults {
 
   RealmObjectMetadata get metadata => _metadata!;
 
-  static RealmResults<T> create<T extends Object?>(
-    RealmResultsHandle handle,
-    Realm realm,
-    RealmObjectMetadata? metadata,
-    [int skip = 0]
-  ) =>
+  static RealmResults<T> create<T extends Object?>(RealmResultsHandle handle, Realm realm, RealmObjectMetadata? metadata, [int skip = 0]) =>
       RealmResults<T>._(handle, realm, metadata, skip);
 }
 
@@ -256,4 +345,21 @@ class _RealmResultsIterator<T extends Object?> implements Iterator<T> {
     }
     return true;
   }
+}
+
+///
+/// Behavior when waiting for subscribed objects to be synchronized/downloaded.
+///
+enum WaitForSyncMode {
+  /// Waits until the objects have been downloaded from the server
+  /// the first time the subscription is created. If the subscription
+  /// already exists, the [RealmResults] is returned immediately.
+  firstTime,
+
+  /// Always waits until the objects have been downloaded from the server.
+  always,
+
+  /// Never waits for the download to complete, but keeps downloading the
+  /// objects in the background.
+  never,
 }